Move lifecycles to their own folder
