+++ server/bootstrap/index.js
@@ -1,11 +1,15 @@
 'use strict';
 
+const beforeUpdateLifecycle = require( '../lifecycles/before-update' );
 const transformApiResponseMiddleware = require( '../middlewares/transform-api-response' );
-const lifecycleBeforeUpdate = require( './lifecycle-before-update' );
 const schemaValidation = require( './schema-validation' );
 
 module.exports = async params => {
-  await transformApiResponseMiddleware( params );
   await schemaValidation( params );
-  await lifecycleBeforeUpdate( params );
+
+  // Middlewares.
+  await transformApiResponseMiddleware( params );
+
+  // Lifecycles.
+  await beforeUpdateLifecycle( params );
 };
